requirements.txt
streamlit
google-generativeai
pypdf
plotly